Updated again

+++ Entities/Product.cs
@@ -8,7 +8,7 @@ public class Product
     [Key]
     public int Id {get; set;}
     public string? Title { get; set; }
-    public ulong Quantity { get; set; }
+    public int Quantity { get; set; }
     public double Price { get; set; }
 
 }
