+++ Services/CalculateService.cs
@@ -2,7 +2,7 @@ namespace Task1.Services;
 
 public class CalculateService
 {
-    public double Calculate(double vat, ulong amount, double price)
+    public double Calculate(double vat, int amount, double price)
     {
         var calculate = ((amount * price) * (1 + vat));
